docs/post_server/requirements.txt
grpcio==1.66.1
grpcio-tools==1.66.1

uuid
PyJWT==2.6.0
cryptography==38.0.4

Flask==2.0.3
Flask-SQLAlchemy==2.5.1
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.3.1
marshmallow==3.14.1
psycopg2-binary
python-dotenv==0.19.2
Werkzeug==2.2.2
SQLAlchemy
pytest==7.4.0
pytest-cov==4.1.0

uvicorn
kafka-python
pydantic
confluent_kafka